Fix trigger buttons turning black on large bombs due to scaling issue
The scale of the trigger buttons on the prefab are very small. Due to this, when the module is scaled to fit on a bomb it shrinks them even more causing them to lose their texture and become purely black. This fixes the issue by scaling the trigger buttons models on the prefab by 10, and decreasing the scale back to fit with the import scale of each model instead.